Move all components into a components directory
