+++ src/components/Checkbox/style.less
