+++ src/components/Rate/index.js
