+++ src/components/Rate/style.less
