+++ src/components/Slider/style.less
